feat: app - server 분리
