+++ src/routes/index.js
@@ -1,0 +1,5 @@
+const express = require("express");
+
+const routes = express.Router();
+// try-catch문에 대해서 err catch 해주는 부분이 라우터 단에서 필요하다.
+module.exports = { routes };
